Add test-release script & makefile target
Running the script sends the already built package to
test pypi, then uses pip to download and install it from
there, and does a minimal invocation of the installed
package. If the package has not been built correctly,
this might detect some errors before we send it to
the real pypi.