Add --extra-index-url to test-release scripts
This would be needed if the project had any dependencies.
It does not, so is redundant, but I figured better to add
it now while I just read about it, and then when dependencies
are added later (or when this gets cut and paste into a
different project), then hopefully things should just work.