README suggests not using Colorama to print colors
Colorama's primary purpose is to convert ANSI codes
to work on Windows. People are forever requesting or
submitting ways to improve the capabilities of Colorama's
Fore, Back, Style constants. I don't want to bloat out the
library with making fancy ways to print colors, when it's
never going to be as good as other libraries like Blessings
or Rich that are dedicated soley to that. Instead we aim
to do one thing well, i.e. convert ansi codes, whether you
generate them using Colorama, Blessings or manually,
into win32 calls so they work on Windows as well.